[FEATURE] Take unknown/invalid extension versions into account
A new argument "--unknown-extension-version-action" can be used to control
the behaviour of the check script if TYPO3 reports an extension with an
invalid version string.

Valid values for this argument are: "unknown", "show" and "ignore". The
Option "unknown" is the default.

This feature is based on a pull request by Markus Gerdes (mgerdes-atl).