[BUGFIX] Fix parameter validation
The parameter --unknown-extension-version-action lets users control how
unknown versions should be treated. The value "ignore" supresses the
display of the extension.

Before this fix, the script did not recognize the value "ignore" due to
an inproper validation. This patch fixes the validation function.

Thanks to a2yp for providing the patch and submitting the pull request.